+++ changelog.md
@@ -1,7 +1,7 @@
 # Change Log
 
-## `master`
-- update common dependency to 0.2.3
+## `v0.1.1`
+- update common dependency to 0.2.4
 - provide more feedback when sending using testhub
 - retry send upon server-busy
 - use a new connection for each sender and receiver